+++ src/csvm/csvm_frequency_matrix.cc
@@ -29,11 +29,11 @@ void FreqMatrix::setZeros(){
  
 }
 
-void FreqMatrix::addCombo(int wordA, int wordB, float activationA, float activationB){
+void FreqMatrix::addCombo(int wordA, int wordB, double activationA, double activationB){
    triangle[getTriangleIndex(wordA, wordB)] += activationA * activationB;
 }
 
-float FreqMatrix::getCombo(int wordA, int wordB){
+double FreqMatrix::getCombo(int wordA, int wordB){
    return triangle[getTriangleIndex(wordA, wordB)];
 }
